+++ packages/api/src/test/EmployeeApiTest.ts
@@ -33,9 +33,13 @@ export namespace EmployeeApiTest {
     | "timeseries"
     | "mediaReference"
     | "geotimeSeriesReference"
-    | "lastClockIn"
-    | "dateOfBirth"
-    | "isActive";
+    | "isActive"
+    | "yearsOfExperience"
+    | "rank"
+    | "performanceScore"
+    | "hourlyRate"
+    | "dateOfJoining"
+    | "lastUpdated";
 
   export interface Links {
     readonly lead: $SingleLinkAccessor<EmployeeApiTest>;
@@ -54,8 +58,12 @@ export namespace EmployeeApiTest {
       | $PropType["geotimeSeriesReference"]
       | undefined;
     readonly isActive: $PropType["boolean"] | undefined;
-    readonly lastClockIn: $PropType["timestamp"] | undefined;
-    readonly dateOfBirth: $PropType["datetime"] | undefined;
+    readonly yearsOfExperience: $PropType["long"] | undefined;
+    readonly rank: $PropType["short"] | undefined;
+    readonly performanceScore: $PropType["double"] | undefined;
+    readonly hourlyRate: $PropType["float"] | undefined;
+    readonly dateOfJoining: $PropType["datetime"] | undefined;
+    readonly lastUpdated: $PropType["timestamp"] | undefined;
   }
   export type StrictProps = Props;
 
@@ -112,8 +120,12 @@ export interface EmployeeApiTest extends $ObjectTypeDefinition {
         "single"
       >;
       isActive: $PropertyDef<"boolean", "nullable", "single">;
-      lastClockIn: $PropertyDef<"timestamp", "nullable", "single">;
-      dateOfBirth: $PropertyDef<"datetime", "nullable", "single">;
+      yearsOfExperience: $PropertyDef<"long", "nullable", "single">;
+      rank: $PropertyDef<"short", "nullable", "single">;
+      performanceScore: $PropertyDef<"double", "nullable", "single">;
+      hourlyRate: $PropertyDef<"float", "nullable", "single">;
+      dateOfJoining: $PropertyDef<"datetime", "nullable", "single">;
+      lastUpdated: $PropertyDef<"timestamp", "nullable", "single">;
     };
     rid: "ri.ontology.main.object-type.401ac022-89eb-4591-8b7e-0a912b9efb44";
     status: "ACTIVE";