+++ packages/client/src/createClient.ts
@@ -17,8 +17,8 @@
 import type {
   ActionDefinition,
   FetchPageArgs,
-  FilteredPropertyKeys,
   InterfaceDefinition,
+  Logger,
   NullabilityAdherence,
   ObjectOrInterfaceDefinition,
   ObjectSet,
@@ -48,7 +48,6 @@ import { applyAction } from "./actions/applyAction.js";
 import { additionalContext, type Client } from "./Client.js";
 import { createMinimalClient } from "./createMinimalClient.js";
 import { fetchMetadataInternal } from "./fetchMetadata.js";
-import { type Logger } from "./logger/Logger.js";
 import { MinimalLogger } from "./logger/MinimalLogger.js";
 import type { MinimalClient } from "./MinimalClientContext.js";
 import { fetchPage } from "./object/fetchPage.js";
@@ -196,7 +195,7 @@ export function createClientInternal(
           return {
             createMediaReference: async <
               Q extends ObjectTypeDefinition,
-              const L extends FilteredPropertyKeys<Q, "mediaReference">,
+              const L extends PropertyKeys.Filtered<Q, "mediaReference">,
             >(args: {
               data: Blob;
               fileName: string;