+++ packages/client.unstable/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/client.unstable
 
+## 2.2.0-beta.18
+
+### Minor Changes
+
+- 6ac8f0b: api bump
+
+## 2.2.0-beta.17
+
+### Minor Changes
+
+- 6eeb423: Simple create and modify actions for interfaces
+- 37f8e60: api bump for concrete actions
+
 ## 2.2.0-beta.16
 
 ### Minor Changes