+++ packages/generator-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/generator-utils
 
+## 2.2.0-beta.18
+
+## 2.2.0-beta.17
+
 ## 2.2.0-beta.16
 
 ## 2.2.0-beta.15