+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @osdk/generator
 
+## 2.2.0-beta.18
+
+### Patch Changes
+
+- Updated dependencies [56aa502]
+  - @osdk/api@2.2.0-beta.18
+  - @osdk/generator-converters@2.2.0-beta.18
+
+## 2.2.0-beta.17
+
+### Minor Changes
+
+- b61e9f5: Make Query Map Types Optional
+- 430dee3: JSdoc for object properties are generated
+- 67e772e: Fixes action parameter generation to enforce strict fields
+- 791e655: Add support for geoshape and geohash action parameters.
+
+### Patch Changes
+
+- Updated dependencies [c6bee50]
+- Updated dependencies [abe4897]
+- Updated dependencies [70b4464]
+- Updated dependencies [791e655]
+  - @osdk/api@2.2.0-beta.17
+  - @osdk/generator-converters@2.2.0-beta.17
+
 ## 2.2.0-beta.16
 
 ### Patch Changes